+++ addons/web_editor/static/src/js/wysiwyg/widgets/link_dialog.js
@@ -1,150 +1,27 @@
 odoo.define('wysiwyg.widgets.LinkDialog', function (require) {
 'use strict';
 
-var core = require('web.core');
-var Dialog = require('wysiwyg.widgets.Dialog');
-const OdooEditorLib = require('web_editor.odoo-editor');
-const wysiwygUtils = require('@web_editor/js/wysiwyg/wysiwyg_utils');
+const Dialog = require('wysiwyg.widgets.Dialog');
+const Link = require('wysiwyg.widgets.Link');
 
-var _t = core._t;
 
-/**
- * Allows to customize link content and style.
- */
-var LinkDialog = Dialog.extend({
+// This widget is there only to extend Link and be instantiated by LinkDialog.
+const _DialogLinkWidget = Link.extend({
     template: 'wysiwyg.widgets.link',
-    xmlDependencies: (Dialog.prototype.xmlDependencies || []).concat([
-        '/web_editor/static/src/xml/wysiwyg.xml'
-    ]),
-    events: _.extend({}, Dialog.prototype.events || {}, {
-        'input': '_onAnyChange',
+    events: _.extend({}, Link.prototype.events || {}, {
         'change [name="link_style_color"]': '_onTypeChange',
-        'change': '_onAnyChange',
-        'input input[name="url"]': '_onURLInput',
     }),
 
-    /**
-     * @constructor
-     * @param {Boolean} linkInfo.isButton - whether if the target is a button element.
-     */
-    init: function (parent, options, editable, data, link) {
-        this.options = options || {};
-        this._super(parent, _.extend({
-            title: _t("Link to"),
-        }, this.options));
-
-        this.data = data || {};
-        this.isButton = this.data.isButton;
-        // Using explicit type 'link' to preserve style when the target is <button class="...btn-link"/>.
-        this.colorsData = [
-            {type: this.isButton ? 'link' : '', label: _t("Link"), btnPreview: 'link'},
-            {type: 'primary', label: _t("Primary"), btnPreview: 'primary'},
-            {type: 'secondary', label: _t("Secondary"), btnPreview: 'secondary'},
-            // Note: by compatibility the dialog should be able to remove old
-            // colors that were suggested like the BS status colors or the
-            // alpha -> epsilon classes. This is currently done by removing
-            // all btn-* classes anyway.
-        ];
-
-        this.editable = editable;
-        this.$editable = $(editable);
-        this.data = data || {};
-
-        this.data.className = this.data.className || "";
-        this.data.iniClassName = this.data.iniClassName || "";
-
-        if (link) {
-            const range = document.createRange();
-            range.selectNodeContents(link);
-            this.data.range = range;
-        } else {
-            const selection = editable && editable.ownerDocument.getSelection();
-            this.data.range = selection && selection.rangeCount && selection.getRangeAt(0);
-        }
-
-        if (this.data.range) {
-            this.needLabel = this.data.range.collapsed;
-            let $link;
-            if (link) {
-                $link = $(link);
-            } else {
-                const firstLinkInSelection = OdooEditorLib.getInSelection(this.editable.ownerDocument, 'a');
-                $link = $(firstLinkInSelection);
-            }
-            this.data.iniClassName = $link.attr("class") || "";
-            this.colorCombinationClass = false;
-            let $node = $link;
-            while ($node.length && !$node.is('body')) {
-                const className = $node.attr('class') || '';
-                const m = className.match(/\b(o_cc\d+)\b/g);
-                if (m) {
-                    this.colorCombinationClass = m[0];
-                    break;
-                }
-                $node = $node.parent();
-            }
-            this.data.className = this.data.iniClassName.replace(/(^|\s+)btn(-[a-z0-9_-]*)?/gi, ' ');
-
-            let textLink = link;
-            if (!textLink) {
-                const startLink = $(this.data.range.startContainer).closest('a')[0];
-                const endLink = $(this.data.range.endContainer).closest('a')[0];
-                const isLink = startLink && (startLink === endLink);
-                textLink = isLink ? startLink : this.data.range.cloneContents();
-            }
-            const [encodedText, images] = wysiwygUtils.encodeNodeToText(textLink);
-            this.data.originalText = wysiwygUtils.decodeText(encodedText, images);
-            this.data.content = encodedText;
-            this.data.images = images;
-            this.data.url = $link.attr('href') || '';
-        } else {
-            this.data.content = this.data.content ? this.data.content.replace(/[ \t\r\n]+/g, ' ') : '';
-            this.needLabel = true;
-        }
-
-        var allBtnClassSuffixes = /(^|\s+)btn(-[a-z0-9_-]*)?/gi;
-        var allBtnShapes = /\s*(rounded-circle|flat)\s*/gi;
-        this.data.className = this.data.iniClassName
-            .replace(allBtnClassSuffixes, ' ')
-            .replace(allBtnShapes, ' ');
-        // 'o_submit' class will force anchor to be handled as a button in linkdialog.
-        if (/(?:s_website_form_send|o_submit)/.test(this.data.className)) {
-            this.isButton = true;
-        }
-    },
     /**
      * @override
      */
     start: function () {
         this.buttonOptsCollapseEl = this.el.querySelector('#o_link_dialog_button_opts_collapse');
-
         this.$styleInputs = this.$('input.link-style');
         this.$styleInputs.prop('checked', false).filter('[value=""]').prop('checked', true);
-        if (this.data.iniClassName) {
-            _.each(this.$('input[name="link_style_color"], select[name="link_style_size"] > option, select[name="link_style_shape"] > option'), el => {
-                var $option = $(el);
-                if ($option.val() && this.data.iniClassName.match(new RegExp('(^|btn-| |btn-outline-)' + $option.val()))) {
-                    if ($option.is("input")) {
-                        $option.prop("checked", true);
-                    } else {
-                        $option.parent().find('option').removeAttr('selected').removeProp('selected');
-                        $option.parent().val($option.val());
-                        $option.attr('selected', 'selected').prop('selected', 'selected');
-                    }
-                }
-            });
+        if (this.data.isNewWindow) {
+            this.$('we-button.o_we_checkbox_wrapper').toggleClass('active', true);
         }
-        if (this.data.url) {
-            var match = /mailto:(.+)/.exec(this.data.url);
-            this.$('input[name="url"]').val(match ? match[1] : this.data.url);
-            this._onURLInput();
-        }
-
-        this._updateOptionsUI();
-        this._adaptPreview();
-
-        this.$('input:visible:first').focus();
-
         return this._super.apply(this, arguments);
     },
 
@@ -168,7 +45,7 @@ var LinkDialog = Dialog.extend({
         var allWhitespace = /\s+/gi;
         var allStartAndEndSpace = /^\s+|\s+$/gi;
         var allBtnTypes = /(^|[ ])(btn-secondary|btn-success|btn-primary|btn-info|btn-warning|btn-danger)([ ]|$)/gi;
-        this.data.className = data.classes.replace(allWhitespace, ' ').replace(allStartAndEndSpace, '');
+        this.data.classes = data.classes.replace(allWhitespace, ' ').replace(allStartAndEndSpace, '');
         if (data.classes.replace(allBtnTypes, ' ')) {
             this.data.style = {
                 'background-color': '',
@@ -177,11 +54,6 @@ var LinkDialog = Dialog.extend({
         }
         this.data.isNewWindow = data.isNewWindow;
         this.final_data = this.data;
-        const returnValue = this._super.apply(this, arguments);
-        if (this.insertedAnchors && this.insertedAnchors.length) {
-            $(this.insertedAnchors[this.insertedAnchors.length - 1]).focus();
-        }
-        return returnValue;
     },
 
     //--------------------------------------------------------------------------
@@ -189,9 +61,7 @@ var LinkDialog = Dialog.extend({
     //--------------------------------------------------------------------------
 
     /**
-     * Adapt the link preview to changes.
-     *
-     * @private
+     * @override
      */
     _adaptPreview: function () {
         var data = this._getData();
@@ -205,55 +75,40 @@ var LinkDialog = Dialog.extend({
         };
         this.$("#link-preview").attr(attrs).html((data.content && data.content.length) ? data.content : data.url);
     },
-
     /**
-     * Get the link's data (url, content and styles).
-     *
-     * @private
-     * @returns {Object} {content: String, url: String, classes: String, isNewWindow: Boolean}
+     * @override
      */
-    _getData: function () {
-        var $url = this.$('input[name="url"]');
-        var url = $url.val();
-        var content = this.$('input[name="label"]').val() || url;
-        content = wysiwygUtils.decodeText(content, this.data.images);
-
-        if (!this.isButton && $url.prop('required') && (!url || !$url[0].checkValidity())) {
-            return null;
-        }
-
-        const type = this.$('input[name="link_style_color"]:checked').val() || '';
-        const size = this.$('select[name="link_style_size"]').val() || '';
-        const shape = this.$('select[name="link_style_shape"]').val() || '';
-        const shapes = shape ? shape.split(',') : [];
-        const style = ['outline', 'fill'].includes(shapes[0]) ? `${shapes[0]}-` : '';
-        const shapeClasses = shapes.slice(style ? 1 : 0).join(' ');
-        const classes = (this.data.className || '') +
-            (type ? (` btn btn-${style}${type}`) : '') +
-            (shapeClasses ? (` ${shapeClasses}`) : '') +
-            (size ? (' btn-' + size) : '');
-        var isNewWindow = this.options.forceNewWindow ?
-            this._isFromAnotherHostName(url) : this.$('input[name="is_new_window"]').prop('checked');
-        if (url.indexOf('@') >= 0 && url.indexOf('mailto:') < 0 && !url.match(/^http[s]?/i)) {
-            url = ('mailto:' + url);
-        } else if (url.indexOf(location.origin) === 0 && this.$('#o_link_dialog_url_strip_domain').prop("checked")) {
-            url = url.slice(location.origin.length);
-        }
-        var allWhitespace = /\s+/gi;
-        var allStartAndEndSpace = /^\s+|\s+$/gi;
-        return {
-            content: content,
-            url: this._correctLink(url),
-            classes: classes.replace(allWhitespace, ' ').replace(allStartAndEndSpace, ''),
-            isNewWindow: isNewWindow,
-        };
+    _doStripDomain: function () {
+        return this.$('#o_link_dialog_url_strip_domain').prop('checked');
     },
     /**
-     * @private
+     * @override
      */
-    _updateOptionsUI: function () {
-        const el = this.el.querySelector('[name="link_style_color"]:checked');
-        $(this.buttonOptsCollapseEl).collapse(el && el.value ? 'show' : 'hide');
+    _getLinkOptions: function () {
+        const options = [
+            'input[name="link_style_color"]',
+            'select[name="link_style_size"] > option',
+            'select[name="link_style_shape"] > option',
+        ];
+        return this.$(options.join(','));
+    },
+    /**
+     * @override
+     */
+    _getLinkShape: function () {
+        return this.$('select[name="link_style_shape"]').val() || '';
+    },
+    /**
+     * @override
+     */
+    _getLinkSize: function () {
+        return this.$('select[name="link_style_size"]').val() || '';
+    },
+    /**
+     * @override
+     */
+    _getLinkType: function () {
+        return this.$('input[name="link_style_color"]:checked').val() || '';
     },
     /**
      * @private
@@ -271,18 +126,33 @@ var LinkDialog = Dialog.extend({
         }
     },
     /**
-     * @private
+     * @override
      */
-    _correctLink: function (url) {
-        if (url.indexOf('mailto:') === 0 || url.indexOf('tel:') === 0) {
-            url = url.replace(/^tel:([0-9]+)$/, 'tel://$1');
-        } else if (url.indexOf('@') !== -1 && url.indexOf(':') === -1) {
-            url = 'mailto:' + url;
-        } else if (url.indexOf('://') === -1 && url[0] !== '/'
-                    && url[0] !== '#' && url.slice(0, 2) !== '${') {
-            url = 'http://' + url;
+    _isNewWindow: function (url) {
+        if (this.options.forceNewWindow) {
+            return this._isFromAnotherHostName(url);
+        } else {
+            return this.$('input[name="is_new_window"]').prop('checked');
         }
-        return url;
+    },
+    /**
+     * @override
+     */
+    _setSelectOption: function ($option, active) {
+        if ($option.is("input")) {
+            $option.prop("checked", active);
+        } else if (active) {
+            $option.parent().find('option').removeAttr('selected').removeProp('selected');
+            $option.parent().val($option.val());
+            $option.attr('selected', 'selected').prop('selected', 'selected');
+        }
+    },
+    /**
+     * @override
+     */
+    _updateOptionsUI: function () {
+        const el = this.el.querySelector('[name="link_style_color"]:checked');
+        $(this.buttonOptsCollapseEl).collapse(el && el.value ? 'show' : 'hide');
     },
 
     //--------------------------------------------------------------------------
@@ -292,25 +162,43 @@ var LinkDialog = Dialog.extend({
     /**
      * @private
      */
-    _onAnyChange: function () {
-        this._adaptPreview();
-    },
-    /**
-     * @private
-     */
     _onTypeChange() {
         this._updateOptionsUI();
     },
     /**
-     * @private
+     * @override
      */
     _onURLInput: function () {
-        var $linkUrlInput = this.$('#o_link_dialog_url_input');
-        $linkUrlInput.closest('.form-group').removeClass('o_has_error').find('.form-control, .custom-select').removeClass('is-invalid');
-        let value = $linkUrlInput.val();
-        let isLink = value.indexOf('@') < 0;
-        this.$('input[name="is_new_window"]').closest('.form-group').toggleClass('d-none', !isLink);
-        this.$('.o_strip_domain').toggleClass('d-none', value.indexOf(window.location.origin) !== 0);
+        this._super(...arguments);
+        this.$('#o_link_dialog_url_input').closest('.form-group').removeClass('o_has_error').find('.form-control, .custom-select').removeClass('is-invalid');
+    },
+});
+
+/**
+ * Allows to customize link content and style.
+ */
+ const LinkDialog = Dialog.extend({
+    init: function (parent, ...args) {
+        this._super(...arguments);
+        this.linkWidget = new _DialogLinkWidget(this, ...args);
+    },
+    start: async function () {
+        const res = await this._super(...arguments);
+        this.linkWidget.appendTo(this.$el);
+        return res;
+    },
+
+    //--------------------------------------------------------------------------
+    // Public
+    //--------------------------------------------------------------------------
+
+    /**
+     * @override
+     */
+    save: function () {
+        this.linkWidget.save();
+        this.final_data = this.linkWidget.final_data;
+        return this._super(...arguments);
     },
 });
